mapGeneration/requirements.txt
Flask
matplotlib
mysql-connector-python
PyJWT
jwt